% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossCompareMarkers.R
\name{crossCompareMarkers}
\alias{crossCompareMarkers}
\title{crossCompareMarkers}
\usage{
crossCompareMarkers(result1, result2, marker_genes, corr_threshold = 0.5)
}
\arguments{
\item{result1}{Output potential marker list from findMechanismMarkers() for comparison 1.}

\item{result2}{Output potnetial marker list from findMechanismMarkers() for comparison 2.}

\item{marker_genes}{Character vector. Marker genes used in both comparisons.}

\item{corr_threshold}{Numeric. Absolute R threshold for selecting strongly correlated genes. Default is 0.5.}
}
\value{
A list containing:
\itemize{
\item potential_markers_1: character vector of candidate marker genes from comparison 1
\item potential_markers_2: character vector of candidate marker genes from comparison 2
\item shared_potential_markers: intersected genes from the two potential marker lists
}
}
\description{
This function takes the output of two findMechanismMarkers() runs and identifies shared genes that are downregulated and strongly correlated in both comparisons.
}
\details{
Cross compare marker gene lists from two comparision and identify promising markers
}
