% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMechanismMarkers.R
\name{findMechanismMarkers}
\alias{findMechanismMarkers}
\title{findMechanismMarkers}
\usage{
findMechanismMarkers(
  seurat_obj,
  ident.1,
  ident.2,
  top_n_DEGs_volcano = 20,
  top_n_DEGs_pcc = 500,
  direction_DEGs = NA,
  logfc.threshold = 0,
  pval.threshold = 0.05,
  marker_genes,
  corr_threshold = 0.5
)
}
\arguments{
\item{seurat_obj}{A labeled Seurat object. The Seurat object returned from isletAutoLabeler() is ready for use.}

\item{ident.1}{Character. Cell identity label for population 1.}

\item{ident.2}{Character. Cell identity label for population 2.}

\item{top_n_DEGs_volcano}{Integer. Number of DEGs to visualize in the volcano plot (default = 20).}

\item{top_n_DEGs_pcc}{Integer. Number of top DE genes to consider in PCC (default = 500).}

\item{direction_DEGs}{Character. "up" or "down" to select direction of DE genes (default = NA, no direction filtering).}

\item{logfc.threshold}{Numeric. Log2FC threshold for DE filtering (default = 0).}

\item{pval.threshold}{Numeric. p-value threshold for DE filtering (default = 0.05).}

\item{marker_genes}{Character vector. Marker genes to calculate PCC against (input required).}

\item{corr_threshold}{Numeric. Absolute R threshold for selecting strongly correlated genes (default = 0.5).}
}
\value{
A list containing:
\itemize{
\item top_DEGs: DEGs used for volcano plot (not filtered by direction)
\item top_DEGs_filtered: DEGs used for PCC (filtered by direction and LogFC thresholds)
\item volcano_plot: ggplot object of volcano plot
\item cor_matrix_raw: full PCC matrix (raw, unfiltered)
\item cor_matrix_filtered: PCC matrix (cells expressing both genes)
\item strongly_correlated_genes: genes with abs(R) > threshold using raw data
\item strongly_correlated_genes_filtered: genes with abs(R) > threshold using filtered data
\item markerwise_correlation_tables_raw: correlation tables per marker (raw)
\item combined_table_raw: gene-wise PCC with all marker genes (raw)
\item markerwise_correlation_tables_filtered: correlation tables per marker
\item combined_table_filtered: gene-wise PCC with all marker genes
}
}
\description{
Performs differential expression (DE) and Pearson correlation coefficient (PCC) analysis to identify potential marker genes associated with a user-defined cellular mechanism (e.g., maturation, proliferation) across two populations.
}
\details{
Find potential markers using DE and PCC analysis
}
