% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isletAutoLabeler code.R
\name{isletAutoLabeler}
\alias{isletAutoLabeler}
\title{Auto-label islet cell types}
\usage{
isletAutoLabeler(
  seurat_obj,
  assay = "integrated",
  cluster_col = "seurat_clusters",
  source = "Human"
)
}
\arguments{
\item{seurat_obj}{A Seurat object to be labeled.}

\item{assay}{The assay to use (default = "integrated")}

\item{cluster_col}{Column in metadata to use for clustering (default = "seurat_clusters")}

\item{source}{Cell source, used as a prefix for the cell type label (e.g., "Human", "SC")}
}
\value{
Updated Seurat object with a new column named "CellType".
}
\description{
Description: Automatically labels Alpha, Beta, Bihormonal, and Other cell clusters in a Seurat object using average marker expression.
Example Usage: isletAutoLabeler(seurat_obj, assay = "RNA", cluster_col = "unintegrated_clusters", source = "SC")
}
\details{
Author: Traye Lin
Date: 2025-04-30
}
